Data persistence across scenes
The scene was duplicated and a portal was added to travel to the other scene. An interface was added for the file handler so other types can be supported easier.  The ref keyword was removed because of manipulation. And scene data is read from memory and loaded across scenes rather than through IO operations.